+++ backend/canisters/group/impl/src/timer_job_types.rs
@@ -4,7 +4,7 @@ use chat_events::MessageContentInternal;
 use ledger_utils::process_transaction;
 use serde::{Deserialize, Serialize};
 use tracing::error;
-use types::{BlobReference, CanisterId, MessageId, MessageIndex, PendingCryptoTransaction, TransactionId, UserId};
+use types::{BlobReference, CanisterId, MessageId, MessageIndex, PendingCryptoTransaction, UserId};
 use utils::consts::MEMO_PRIZE_REFUND;
 use utils::time::{MINUTE_IN_MS, SECOND_IN_MS};
 
@@ -57,7 +57,7 @@ pub struct NotifyEscrowCanisterOfDepositJob {
     pub swap_id: u32,
     pub thread_root_message_index: Option<MessageIndex>,
     pub message_id: MessageId,
-    pub transaction_id: TransactionId,
+    pub transaction_index: u64,
     pub attempt: u32,
 }
 
@@ -67,14 +67,14 @@ impl NotifyEscrowCanisterOfDepositJob {
         swap_id: u32,
         thread_root_message_index: Option<MessageIndex>,
         message_id: MessageId,
-        transaction_id: TransactionId,
+        transaction_index: u64,
     ) {
         let job = NotifyEscrowCanisterOfDepositJob {
             user_id,
             swap_id,
             thread_root_message_index,
             message_id,
-            transaction_id,
+            transaction_index,
             attempt: 0,
         };
         job.execute();
@@ -241,7 +241,7 @@ impl Job for NotifyEscrowCanisterOfDepositJob {
                             self.user_id,
                             self.thread_root_message_index,
                             self.message_id,
-                            self.transaction_id,
+                            self.transaction_index,
                             state.env.now(),
                         );
                     });
@@ -263,7 +263,7 @@ impl Job for NotifyEscrowCanisterOfDepositJob {
                                 user_id: self.user_id,
                                 thread_root_message_index: self.thread_root_message_index,
                                 message_id: self.message_id,
-                                transaction_id: self.transaction_id,
+                                transaction_index: self.transaction_index,
                                 attempt: self.attempt + 1,
                             }),
                             now + 10 * SECOND_IN_MS,