+++ backend/canister_installer/src/lib.rs
@@ -57,11 +57,12 @@ async fn install_service_canisters_impl(
 
     let user_index_canister_wasm = get_canister_wasm(CanisterName::UserIndex, version);
     let user_index_init_args = user_index_canister::init::Args {
-        service_principals: vec![principal],
+        governance_principals: vec![principal],
         user_canister_wasm: CanisterWasm::default(),
         local_user_index_canister_wasm: CanisterWasm::default(),
         group_index_canister_id: canister_ids.group_index,
         notifications_index_canister_id: canister_ids.notifications_index,
+        identity_canister_id: canister_ids.identity,
         proposals_bot_canister_id: canister_ids.proposals_bot,
         storage_index_canister_id: canister_ids.storage_index,
         cycles_dispenser_canister_id: canister_ids.cycles_dispenser,
@@ -74,7 +75,7 @@ async fn install_service_canisters_impl(
 
     let group_index_canister_wasm = get_canister_wasm(CanisterName::GroupIndex, version);
     let group_index_init_args = group_index_canister::init::Args {
-        service_principals: vec![principal],
+        governance_principals: vec![principal],
         group_canister_wasm: CanisterWasm::default(),
         community_canister_wasm: CanisterWasm::default(),
         local_group_index_canister_wasm: CanisterWasm::default(),
@@ -88,7 +89,7 @@ async fn install_service_canisters_impl(
 
     let notifications_index_canister_wasm = get_canister_wasm(CanisterName::NotificationsIndex, version);
     let notifications_index_init_args = notifications_index_canister::init::Args {
-        service_principals: vec![principal],
+        governance_principals: vec![principal],
         push_service_principals: vec![principal],
         user_index_canister_id: canister_ids.user_index,
         authorizers: vec![canister_ids.user_index, canister_ids.group_index],
@@ -100,6 +101,7 @@ async fn install_service_canisters_impl(
 
     let identity_canister_wasm = get_canister_wasm(CanisterName::Identity, version);
     let identity_init_args = identity_canister::init::Args {
+        governance_principals: vec![principal],
         user_index_canister_id: canister_ids.user_index,
         cycles_dispenser_canister_id: canister_ids.cycles_dispenser,
         wasm_version: version,