+++ backend/canisters/user_index/api/src/updates/c2c_update_user_principal.rs
@@ -3,10 +3,12 @@ use serde::{Deserialize, Serialize};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
+    pub old_principal: Principal,
     pub new_principal: Principal,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
     Success,
+    InternalError(String),
 }