+++ backend/canisters/escrow/impl/src/model/notify_status_change_queue.rs
@@ -3,7 +3,6 @@ use std::collections::VecDeque;
 
 #[derive(Serialize, Deserialize, Default)]
 pub struct NotifyStatusChangeQueue {
-    #[serde(default)]
     swaps: VecDeque<u32>,
 }