+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Upgrade Diamond members first ([#5214](https://github.com/open-chat-labs/open-chat/pull/5214))
+
 ## [[2.0.1011](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1011-local_user_index)] - 2024-01-18
 
 ### Added